+++ admin/admin_submissions.php
@@ -181,7 +181,13 @@ EOD;
                 <label for="quickedit_location_nation">Nation</label>
                 <input class="meeting-input" type="text" name="quickedit_location_nation" size="50" id="quickedit_location_nation">
             </div>
-            
+            <fieldset>
+                <legend>Meeting Geolocation (auto calculated)</legend>
+                <label for="quickedit_latitude">Latitude</label>
+                <input class="meeting-input" type="number" name="quickedit_latitude" id="quickedit_latitude" disabled>
+                <label for="quickedit_longitude">Longitude</label>
+                <input class="meeting-input" type="number" name="quickedit_longitude" id="quickedit_longitude" disabled>
+            </fieldset>
         </div>
     </div>
 </div>